add dedicated unittests for exceptions that have a writeTo method (custom serialization)

+++ core/src/main/java/org/elasticsearch/index/engine/RecoveryEngineException.java
@@ -39,7 +39,7 @@ public class RecoveryEngineException extends EngineException {
 
     public RecoveryEngineException(StreamInput in) throws IOException{
         super(in);
-        phase = in.readByte();
+        phase = in.readInt();
     }
 
     @Override
